refactor (pyutils): Convert Miller-Rabin function to a class
The Miller-Rabin algorithm is slow compared to trial division for small
n, so it is necessary for performance reasons to first attempt trial
division versus an intial list of primes.

To avoid repeatedly calculating these initial primes (and to avoid
hard-coding their values in the source code) a class is implemented,
with calculation of initial primes delegated to the __init__ method.

Primality of a given number can then be checked by calling the class'
`is_prime` method.